# contracts/openapi-diet-management-v1.yaml
openapi: 3.1.0
info:
  title: Diet Management API
  version: v1-fixed
  description: >
    스냅샷 계약 (SSOT 본문: docs/requirements/feature-diet-management-api-contract-v1.md).
    구현체는 본 파일과 문서를 함께 유지한다.
servers:
  - url: http://localhost:3000
    description: 로컬 스텁 (apps/server)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, details, traceId]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        traceId: { type: string }
...
    MealInput:
      type: object
      additionalProperties: true
...
    OcrResult:
      type: object
      properties:
        calories: { type: number }
        carbohydrate: { type: number }
        protein: { type: number }
        fat: { type: number }
        confidence: { type: number, minimum: 0, maximum: 1 }
        missingFields:
          type: array
          items: { type: string }
        remainingFreeQuota: { type: integer }
    StatsResponse:
      type: object
      required: [aggregatedAt, isStale, staleHours, timezone, summary]
      properties:
        aggregatedAt: { type: string, format: date-time }
        isStale: { type: boolean }
        staleHours: { type: number }
        timezone: { type: string }
        summary:
          type: object
          properties:
            calories: { type: number }
            carbohydrate: { type: number }
            protein: { type: number }
            fat: { type: number }
    Entitlements:
      type: object
      properties:
        ocrQuotaLimit: { type: integer }
        ocrQuotaUsed: { type: integer }
        ocrPaidEnabled: { type: boolean }
        adFreeEnabled: { type: boolean }
        nextPaywallTrigger:
          type: string
          enum: [none, ocr_remaining_1, ocr_exhausted]
    AdsStatus:
      type: object
      properties:
        showBottomBanner: { type: boolean }
        reason:
          type: string
          enum: [default_free, ad_free_purchased]
...
paths:
  /auth/signup:
    post:
      tags: [Auth]
      summary: 회원가입
      responses:
        '201': { description: Created }
        '422': { content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      responses:
        '200': { description: OK }
        '401': { content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
  /me/profile:
    get:
      tags: [Me]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: OK }
    put:
      tags: [Me]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: OK }
        '422': { content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /me/recommendation/recalculate:
    post:
      tags: [Me]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: OK }
  /meals:
    post:
      tags: [Meals]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MealInput' }
      responses:
        '201': { description: Created }
        '422': { content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      tags: [Meals]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: size
          schema: { type: integer, default: 15 }
      responses:
        '200': { description: OK }
  /meals/{mealId}:
    put:
      tags: [Meals]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: mealId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /meals/{mealId}/deactivate:
    patch:
      tags: [Meals]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: mealId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /nutrition/ocr:
    post:
      tags: [OCR]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OcrResult' }
        '402': { content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '429': { content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '503': { content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /stats:
    get:
      tags: [Stats]
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: range
          required: true
          schema:
            type: string
            enum: [meal, day, week, month]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatsResponse' }
  /me/billing/entitlements:
    get:
      tags: [Billing]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Entitlements' }
  /me/billing/checkout:
    post:
      tags: [Billing]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productType]
              properties:
                productType:
                  type: string
                  enum: [premium_monthly]
      responses:
        '200': { description: OK }
        '402': { content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /me/billing/restore:
    post:
      tags: [Billing]
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: OK }
  /me/ads/status:
    get:
      tags: [Ads]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdsStatus' }
...